+++ api/models/mensagem.go
@@ -1,8 +1,8 @@
 package models
 
 type Mensagem struct {
-	ID        uint64 `json:"msgID"`
-	ChatID    uint64 `json:"chatID"`
+	ID        int    `json:"msgID"`
+	ChatID    int    `json:"chatID"`
 	Autor     int    `json:"autorID"`
 	Conteudo  string `json:"conteudo"`
 	HoraEnvio string `json:"hora_enviada"`
